refactor(composio): move live catalog and probe logic to composio domain
